requirements.txt
gradio>=4.44.0
transformers>=4.44.0
torch>=2.1.0
Pillow>=10.0.0
pydantic>=2.0.0
accelerate>=0.24.0
huggingface_hub>=0.23.0
sentencepiece>=0.1.99
bitsandbytes>=0.43.0
safetensors>=0.4.0
peft>=0.11.0
trl>=0.9.0
datasets>=2.20.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
jsonschema>=4.22.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
tqdm>=4.66.0
